hyperliquid-bot/requirements.txt
hyperliquid-python-sdk
python-dotenv
pandas
numpy
ta
requests
websockets
